requirements.txt
fastmcp
pyyaml
qbittorrent-api
transmission-rpc
anyio
httpx